Fix backoff_time type annotation in concurrent() decorator (#5410)

+++ custom_components/hacs/utils/decorator.py
@@ -15,7 +15,7 @@ if TYPE_CHECKING:
 
 def concurrent(
     concurrenttasks: int = DEFAULT_CONCURRENT_TASKS,
-    backoff_time: int = DEFAULT_CONCURRENT_BACKOFF_TIME,
+    backoff_time: float = DEFAULT_CONCURRENT_BACKOFF_TIME,
 ) -> Coroutine[Any, Any, None]:
     """Return a modified function."""
